Fix jenkins test result access denied problem

+++ tests/functional_test/start_test_meta_parser.sh
@@ -1,15 +1,21 @@
 #!/bin/bash
 
+echo -e "\n======== ${BASH_SOURCE[0]} ========"
 repo="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && while [ ! -d .git ] ; do cd ..; done; pwd )"
+here="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+source $repo/utils/common_header.bash
+cd $repo
 
-$repo/utils/setup_dev_env.sh -m docker_host
+echo "########## Setup Test Env"
+$repo/utils/setup_dev_env.sh -m functional_test,docker_host
+. $repo/utils/env_config.sh
 
-rm -r $repo/dist
-
+echo "########## pi_tester.py"
+if [ -d $repo/dist ]; then
+	rm -r $repo/dist
+fi
 $repo/build.sh pyhcfs
 
+echo "########## pi_tester.py"
 cd $repo/tests/functional_test/TestCases/TestMetaParser/docker
-
-# python prepare.py
-
 python startMetaParserDockerTest.py
